chapter1

+++ ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: inseok <inseok@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/15 10:06:24 by inseok            #+#    #+#             */
-/*   Updated: 2023/03/15 10:09:59 by inseok           ###   ########.fr       */
+/*   Updated: 2023/03/19 10:31:29 by inseok           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
 	ft_memset(s, 0, n);
 }
